Add aggressive caching
I made several attempts at caching with GC using a stack, with
SetFinalizer, or uintptr and unsafe.Pointer, but all these ideas
failed in some way. The Go GC is not advanced enough to deal with
this it seems. Nevertheles it seems my carry-less addition is also
not very efficient. Perhaps adding a carry is a better idea after
all.